requirements.txt
# Dependencias del proyecto Crypto Price Pipeline
# Instalar con: pip install -r requirements.txt

requests==2.32.5          # Llamadas HTTP a las APIs (CoinGecko, DolarAPI)
python-dotenv==1.2.1      # Carga segura de credenciales desde .env
pandas==2.3.3             # Análisis y procesamiento del histórico de precios
